+++ build.gradle
@@ -65,3 +65,7 @@ dependencies {
 tasks.named('test') {
     useJUnitPlatform()
 }
+
+tasks.withType(JavaCompile) {
+    options.compilerArgs += '-parameters'
+}
