+++ src/main/java/org/example/easytable/common/utils/DummyCounter.java
