+++ src/main/java/org/example/easytable/reservation/entity/Reservation.java
@@ -20,6 +20,7 @@ public class Reservation extends BaseEntity {
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id", nullable = false)
+    @Setter
     private Member member;
 
     @ManyToOne(fetch = FetchType.LAZY)
